+++ RoomClean/Controllers/UsuarioController.cs
@@ -1,25 +1,220 @@
-﻿using Microsoft.AspNetCore.Authorization;
+﻿using Domain.DTOS;
+using Domain.Entities;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.IdentityModel.Tokens;
+using RoomClean.Context;
 using RoomClean.Services;
+using System;
+using System.ComponentModel.DataAnnotations;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Text;
+using System.Threading.Tasks;
 
 namespace RoomClean.Controllers
 {
-    [ApiController]
     [Route("[controller]")]
-    [Authorize]
+    [ApiController]
     public class UsuarioController : ControllerBase
     {
-        public readonly IUsuarioService _usuarioService;
-        public UsuarioController(IUsuarioService usuarioService)
+        private readonly IUsuarioService _usuarioService;
+        private readonly IConfiguration _configuration;
+        private readonly ApplicationDBContext _context;
+
+        public UsuarioController(IUsuarioService usuarioService, IConfiguration configuration, ApplicationDBContext context)
         {
             _usuarioService = usuarioService;
+            _configuration = configuration;
+            _context = context;
         }
 
-        //[HttpGet]
-        //public async Task<IActionResult> ObtenerLista()
-        //{
-        //    var result = await _usuarioService.ObtenerUsuarios();
-        //    return Ok(result);
-        //}
+        [HttpGet("list")]
+        [Authorize]
+        public async Task<IActionResult> ObtenerUsuarios()
+        {
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+            var rtoken = Jwt.Validartoken(identity, _context);
+
+            if (!rtoken.success)
+                return BadRequest(new { success = false, message = rtoken.message });
+
+            Usuario usuario = rtoken.result;
+
+            if (usuario.FKRol != 1)
+            {
+                return BadRequest("No tienes permisos para esta accion");
+
+            }
+
+            try
+            {
+                var response = await _usuarioService.ObtenerUsuarios();
+                return Ok(response);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new { message = ex.Message });
+            }
+        }
+
+        [HttpGet("list/{id}")]
+        [Authorize]
+        public async Task<IActionResult> ObtenerUsuario(int id)
+        {
+            try
+            {
+                var response = await _usuarioService.ObtenerUsuario(id);
+                return Ok(response);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new { message = ex.Message });
+            }
+        }
+
+        [HttpPost("create")]
+        [Authorize]
+        public async Task<IActionResult> CrearUsuario([FromBody] UsuarioDto request)
+        {
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+            var rtoken = Jwt.Validartoken(identity, _context);
+
+            if (!rtoken.success)
+                return BadRequest(new { success = false, message = rtoken.message });
+
+            Usuario usuario = rtoken.result;
+
+            if (usuario.FKRol != 1)
+            {
+                return BadRequest("No tienes permisos para esta accion");
+
+            }
+
+            try
+            {
+                var response = await _usuarioService.CrearUsuario(request);
+                return Ok(response);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new { message = ex.Message });
+            }
+        }
+
+        [HttpPut("update/{id}")]
+        [Authorize]
+        public async Task<IActionResult> ActualizarUsuario(int id, [FromBody] UsuarioDto request)
+        {
+
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+            var rtoken = Jwt.Validartoken(identity, _context);
+
+            if (!rtoken.success)
+                return BadRequest(new { success = false, message = rtoken.message });
+
+            Usuario usuario = rtoken.result;
+
+            if (usuario.FKRol != 1)
+            {
+                return BadRequest("No tienes permisos para esta accion");
+
+            }
+
+            try
+            {
+                var response = await _usuarioService.ActualizarUsuario(id, request);
+                return Ok(response);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new { message = ex.Message });
+            }
+        }
+
+        [HttpDelete("delete/{id}")]
+        [Authorize]
+        public async Task<IActionResult> EliminarUsuario(int id)
+        {
+
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+            var rtoken = Jwt.Validartoken(identity, _context);
+
+            if (!rtoken.success)
+                return BadRequest(new { success = false, message = rtoken.message });
+
+            Usuario usuario = rtoken.result;
+
+            if (usuario.FKRol != 1)
+            {
+                return BadRequest("No tienes permisos para esta accion");
+
+            }
+
+            try
+            {
+                var response = await _usuarioService.EliminarUsuario(id);
+                return Ok(response);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new { message = ex.Message });
+            }
+        }
+
+        [HttpPost("login")]
+        public async Task<IActionResult> Login([FromBody] LoginModel model)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            string user = model.Correo.ToString();
+            string hashedPassword = ApplicationDBContext.ComputeSha256Hash(model.Contraseña);
+
+            Usuario usuario = _context.Usuarios.FirstOrDefault(x => x.Correo == user && x.Contraseña == hashedPassword);
+
+            if (usuario == null)
+            {
+                return Unauthorized(new { message = "Usuario o contraseña incorrectos" });
+            }
+
+            var jwt = _configuration.GetSection("Jwt").Get<Jwt>();
+            var claims = new[]
+            {
+                new Claim(JwtRegisteredClaimNames.Sub,jwt.Subject),
+                new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
+                new Claim(JwtRegisteredClaimNames.Iat, DateTime.UtcNow.ToString()),
+                new Claim("id", usuario.Id.ToString()),
+                new Claim("rol", usuario.FKRol.ToString()),
+                new Claim("usuario", usuario.Correo)
+            };
+
+            var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_configuration["Jwt:Key"]));
+            var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
+
+
+            var token = new JwtSecurityToken(
+                issuer: _configuration["Jwt:Issuer"],
+                audience: _configuration["Jwt:Audience"],
+                claims: claims,
+                expires: DateTime.Now.AddMinutes(30),
+                signingCredentials: creds);
+
+            return Ok(new { token = new JwtSecurityTokenHandler().WriteToken(token) });
+        }
+    }
+    public class LoginModel
+    {
+        [Required(ErrorMessage = "No se ha ingresado ningun correo")]
+        public string Correo { get; set; }
+
+        [Required(ErrorMessage = "La contraseña es obligatoria.")]
+        public string Contraseña { get; set; }
     }
 }
+
+
